Moved token data to Collection Page
